feat: clean up unused files
